requirements.txt
streamlit==1.52.1
langchain==1.1.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.1
langchain-google-genai==3.2.0
langchain-huggingface==1.1.0
langchain-text-splitters==1.0.0
python-dotenv==1.2.1
faiss-cpu==1.13.1
pypdf==6.4.0
docx2txt==0.9
sentence-transformers==5.1.2

